TRUNK-381: Moving back test classes to /api.
