+++ api/src/test/java/org/openmrs/api/context/TestUsernameAuthenticationScheme.java
